+++ tartufo/cli.py
@@ -45,7 +45,6 @@ class TartufoCLI(click.MultiCommand):
     name="tartufo",
     context_settings=dict(help_option_names=["-h", "--help"]),
 )
-@click.option("--json/--no-json", help="Output in JSON format.", is_flag=True)
 @click.option(
     "--rules",
     multiple=True,
@@ -61,13 +60,6 @@ class TartufoCLI(click.MultiCommand):
     " search patterns. Only applicable if --rules is also specified.",
 )
 @click.option(
-    "--compact/--no-compact",
-    is_flag=True,
-    default=False,
-    show_default=True,
-    help="Enable reduced output.",
-)
-@click.option(
     "--entropy/--no-entropy",
     is_flag=True,
     default=True,
@@ -77,9 +69,16 @@ class TartufoCLI(click.MultiCommand):
 @click.option(
     "--regex/--no-regex",
     is_flag=True,
-    default=False,
+    default=True,
     show_default=True,
     help="Enable high signal regexes checks.",
+)
+@click.option(
+    "--scan-filenames/--no-scan-filenames",
+    is_flag=True,
+    default=True,
+    show_default=True,
+    help="Check the names of files being scanned as well as their contents.",
 )
 @click.option(
     "-ip",
@@ -100,6 +99,22 @@ class TartufoCLI(click.MultiCommand):
     exclude multiple patterns. If not provided (default), no Git object paths
     are excluded unless effectively excluded via the --include-path-patterns
     option.""",
+)
+@click.option(
+    "-of",
+    "--output-format",
+    type=click.Choice(
+        [
+            types.OutputFormat.Json.value,
+            types.OutputFormat.Compact.value,
+            types.OutputFormat.Text.value,
+        ]
+    ),
+    default="text",
+    help="""Specify the format in which the output needs to be generated
+    `--output-format json/compact/text`. Either `json`, `compact` or `text`
+    can be specified. If not provided (default) the output will be generated
+    in `text` format.""",
 )
 @click.option(
     "-xe",