+++ CHANGELOG.md
@@ -14,6 +14,12 @@ Features:
   thanks to @dclayton-godaddy for showing the way.
 * [#244](https://github.com/godaddy/tartufo/pull/244) - Drops support for 
   `--fetch/--no-fetch` option for local scans
+* [#253](https://github.com/godaddy/tartufo/issues/253) - Drops support for `--json` and `--compact`
+  and consolidates the two options into one `---output-format json/compact/text`
+* [#259](https://github.com/godaddy/tartufo/pull/259) - Adds a new 
+  `--scan-filenames/--no-scan-filenames` flag which allows users to enable or disable file name scanning.
+* [#254](https://github.com/godaddy/tartufo/pull/260) - Changes the default value of 
+  `--regex/--no-regex` to True.
 
 Misc: